refactor: change file path for DuedateStrategy
