+++ TodoItems.Core/DuedateStartegy/LeastCountDuedateStrategy.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace TodoItems.Core
+namespace TodoItems.Core.DuedateStartegy
 {
     public class LeastCountDuedateStrategy : IDuedateStrategy
     {
